+++ src/main/java/com/expensetracker/repository/DatabaseManager.java
@@ -1,4 +1,4 @@
-package expensetracker;
+package com.expensetracker.repository;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
